requirements.txt
ollama==0.5.1
python-dotenv==1.1.1
Requests==2.32.4
streamlit==1.41.1
